requirements.txt
asyncpg==0.30.0
fastapi==0.115.4
loguru==0.7.2
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.2.0
SQLAlchemy==2.0.36
uvicorn==0.32.0
pillow==11.1.0